+++ lib/modules/groceries/groceries_page.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:money_manager/modules/groceries/widgets/form_cart_item_widget.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:money_manager/modules/groceries/stores/cart_item_store.dart';
+import 'package:money_manager/modules/groceries/widgets/form_cart_item_widget.dart';
 import 'package:money_manager/modules/groceries/stores/grocery_item_store.dart';
 import 'package:money_manager/modules/groceries/stores/shopping_cart_store.dart';
 
@@ -22,7 +23,7 @@ class GroceriesPage extends StatelessWidget {
         actions: [
           InkWell(
             onTap: () =>
-                Navigator.pushNamed(context, 'groceries-shopping-cart'),
+                Navigator.pushNamed(context, 'groceries/shopping-cart'),
             child: Stack(alignment: Alignment.center, children: [
               const Padding(
                 padding: EdgeInsets.all(10.0),
@@ -52,9 +53,19 @@ class GroceriesPage extends StatelessWidget {
       ),
       floatingActionButton: FloatingActionButton(
         child: const Icon(Icons.add),
-        tooltip: 'Nuevo producto',
-        onPressed: () => Navigator.pushNamed(context, 'groceries-new-product'),
+        onPressed: () => Navigator.pushNamed(context, 'groceries/new-product'),
       ),
+      // floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
+      // bottomNavigationBar: BottomAppBar(
+      //     shape: const CircularNotchedRectangle(),
+      //     child: Row(
+      //         mainAxisSize: MainAxisSize.max,
+      //         mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      //         children: [
+      //           IconButton(icon: const Icon(Icons.home), onPressed: () {}),
+      //           IconButton(
+      //               icon: const Icon(Icons.shopping_cart), onPressed: () {}),
+      //         ])),
       body: Column(
         children: [
           Padding(
